+++ clientsrc/src/store/index.js
@@ -5,19 +5,14 @@ import Vuex, {
 import Axios from "axios";
 import router from "../router";
 import Song from "../models/song";
+import {api} from "../services/AxiosService";
 import {socketStore} from "./socketStore";
+import {favoriteStore} from "./favoriteStore";
+import {recommendStore} from "./recommendStore";
 
 Vue.use(Vuex);
 
-let baseUrl = location.host.includes("localhost") ?
-  "http://localhost:3000/" :
-  "/";
 
-let api = Axios.create({
-  baseURL: baseUrl + "api",
-  timeout: 3000,
-  withCredentials: true
-});
 
 export default new Vuex.Store({
   state: {
@@ -60,32 +55,9 @@ export default new Vuex.Store({
         console.error(error);
       }
     },
-    async getMusicByQueryFav({
-      commit,
-      dispatch
-    }, query) {
-      try {
-        let res = await api.get("favorites/" + query)
-        commit("setSearchedSongs", res.data);
-      } catch (error) {
-
-      }
-
-    },
-    async getMusicByQueryRec({ commit, dispatch }, query) {
-      try {
-
-        let res = await api.get("recommends/" + query)
-        commit("setSearchedSongs", res.data);
-      } catch (error) {
-
-      }
-
-    },
-    async getMusicByQuery({
-      commit,
-      dispatch
-    }, query) {
+   
+   
+    async getMusicByQuery({commit,dispatch}, query) {
       try {
         let url = "https://itunes.apple.com/search?callback=?&term=" + query;
         // @ts-ignore
@@ -105,74 +77,7 @@ export default new Vuex.Store({
         })
       }
     },
-    async addToFavorites({ commit, dispatch }, newFavorite) {
-      let res = await api.post("favorites", {
-        Song: newFavorite
-      })
-      commit("addFavorite", res.data)
-    },
-    async getFavoritesByEmail({ commit, dispatch }) {
-      let res = await api.get("favorites")
-      commit("setFavorites", res.data)
-    },
-    async deleteFavorite({ commit, dispatch }, id) {
-      let res = await api.delete("favorites/" + id)
-      dispatch("getFavoritesByEmail")
-    },
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-    async recommendTo({ commit, dispatch }, { email, song }) {
-      //  debugger
-      let obj = {
-        receiver: email.value,
-        song: song.Song,
-        sender: song.creatorEmail
-      }
-      let res = await api.post("recommends", obj)
-      dispatch("getTotalForPut")
-    },
-
-    async recommendToFromSearch({ commit, dispatch }, { email, song, creatorEmail }) {
-      //also used when recommending from recommends
-      let obj = {
-        receiver: email.value,
-        song: song,
-        sender: creatorEmail
-      }
-      let res = await api.post("recommends", obj)
-      dispatch("getTotalForPut")
-      dispatch("getRecommends")
-    },
-
-
-
-    async getRecommends({ commit, dispatch }) {
-      let res = await api.get("recommends")
-      commit("setRecommends", res.data)
-    },
-
-    async deleteRec({ commit, dispatch }, id) {
-      let res = await api.delete("recommends/" + id)
-      dispatch("getRecommends")
-    },
-
-
-
-    async getTotalForPut({ commit, dispatch }) {
+   async getTotalForPut({ commit, dispatch }) {
       //1
       let resT = await api.get("profile")
       dispatch("scoreTotalRec", resT.data.totalRecommends)
@@ -190,14 +95,6 @@ export default new Vuex.Store({
 
       }
     },
-
-    like({ commit, dispatch }, { newFavorite, id }) {
-      //debugger
-      dispatch("addToFavorites", newFavorite)
-      dispatch("getPositivesForPut")
-      dispatch("deleteRec", id)
-    },
-
     async getPositivesForPut({ commit, dispatch }) {
       //1
       let resT = await api.get("profile")
@@ -216,24 +113,17 @@ export default new Vuex.Store({
 
       }
     },
-
-    async getScoreVariables({ commit, dispatch }) {
+      async getScoreVariables({ commit, dispatch }) {
       let resT = await api.get("profile")
       // console.log(resT.data)
       //dispatch("scoreTotalRec", resT.data.totalRecommends)
 
 
     }
-
-
-
-
-
-
-  },
+   },
   modules: {
-    // favoritesStore,
-    // recommendsStore,
+    favoriteStore,
+    recommendStore,
     socketStore
   }
 });
